Inline worker_pools.py
